pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sputnik"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "requests>=2.0",
  "PyYAML>=6.0",
  "Pillow==11.1.0",
  "rasterio>=1.4.3",
]

[tool.setuptools.packages.find]
include = ["copernicus_utils*"]
exclude = ["data*", "tests*"]
